Lets move the checkclasspath and serviceProviderBuildItem in AWS2-commons support extension
