+++ extensions-support/aws2/deployment/src/main/java/org/apache/camel/quarkus/component/aws/commons/deployment/AwsCommonsProcessor.java
@@ -16,13 +16,21 @@
  */
 package org.apache.camel.quarkus.component.aws.commons.deployment;
 
+import javax.enterprise.inject.spi.DeploymentException;
+
+import io.quarkus.arc.deployment.BeanRegistrationPhaseBuildItem;
+import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.NativeImageProxyDefinitionBuildItem;
+import io.quarkus.deployment.builditem.nativeimage.ServiceProviderBuildItem;
+import software.amazon.awssdk.http.SdkHttpService;
 
 class AwsCommonsProcessor {
 
     private static final String FEATURE = "camel-aws2-commons";
+
+    private static final String APACHE_HTTP_SERVICE = "software.amazon.awssdk.http.apache.ApacheSdkHttpService";
 
     @BuildStep
     FeatureBuildItem feature() {
@@ -33,6 +41,25 @@ class AwsCommonsProcessor {
     NativeImageProxyDefinitionBuildItem httpProxies() {
         return new NativeImageProxyDefinitionBuildItem("org.apache.http.conn.HttpClientConnectionManager",
                 "org.apache.http.pool.ConnPoolControl", "software.amazon.awssdk.http.apache.internal.conn.Wrapped");
+    }
+
+    @BuildStep(loadsApplicationClasses = true)
+    void client(BeanRegistrationPhaseBuildItem beanRegistrationPhase,
+            BuildProducer<ServiceProviderBuildItem> serviceProvider,
+            BuildProducer<NativeImageProxyDefinitionBuildItem> proxyDefinition) {
+        checkClasspath(APACHE_HTTP_SERVICE, "apache-client");
+
+        serviceProvider.produce(new ServiceProviderBuildItem(SdkHttpService.class.getName(), APACHE_HTTP_SERVICE));
+
+    }
+
+    private void checkClasspath(String className, String dependencyName) {
+        try {
+            Class.forName(className, true, Thread.currentThread().getContextClassLoader());
+        } catch (ClassNotFoundException e) {
+            throw new DeploymentException(
+                    "Missing 'software.amazon.awssdk:" + dependencyName + "' dependency on the classpath");
+        }
     }
 
 }
